komenco: Remove submodule
Currently git submodules are used in order to get the komenco repository
below the Dockerfile which is necessary in order to copy it into the
container during build. This makes some trouble when building and
verifying the containers. So the submodule gets removed and the
komenco source of your choice must be manually placed into the
komenco dir.

Change-Id: Ia73b6deb87bfe46d0dcd137a39660a1c0effa0ba